Adding StatsD logging to Swift.
Documentation, including a list of metrics reported and their semantics,
are in the Admin Guide in a new section, "Reporting Metrics to StatsD".

Here is the rationale for doing a deep integration like this versus only
sending metrics to StatsD in middleware.  It's the only way to report
some internal activities of Swift in a real-time manner. So to have one
way of reporting to StatsD and one place/style of configuration, even
some things (like, say, timing of PUT requests into the proxy-server)
which could be logged via middleware are consistently logged the same
way (deep integration via the logger delegate methods).

When log_statsd_host is configured and the pystatsd module is installed,
get_logger() injects a pystatsd.Client object into the logger as
logger.statsd_client.  Then a set of delegate methods on LogAdapter
either pass through to the pystatsd.Client object or become no-ops. This
allows StatsD logging to look like:
    self.logger.increment('some.metric.here')
and do the right thing in all cases and with no messy conditional logic.

Refactored out _one_audit_pass() method in swift/account/auditor.py and
swift/container/auditor.py.

Fixed some misc. PEP8 violations.

Change-Id: Ie968a9ae8771f59ee7591e2ae11999c44bfe33b2