Added tests for StatsD logging.
Also fixed all the existing tests which it broke.  This involved some
refactoring among test helper code.

Since the latest pystatsd version on PyPi is lagging the git repo (and
is missing both the prefix functionality and timing_since() method), I
went ahead and wrote my own class, swift.common.utils.StatsdClient.  The
interface is the same as pystatsd.Client, but the code was written from
scratch.  It's pretty simple, and the tests I added cover it.  This also
frees Swift from an optional dependency on the pystatsd module, making
it easier to enable.