Add device name to *-replicator.removes for DBs.
[Pulling in pre-merge swift patch from
https://review.openstack.org/#/c/11600/]

To tell when replication for a device has finished, it's important to
know when the replicator is removing objects.  This was previously
handled for the object-replicator
(object-replicator.partition.delete.count.<device> and
object-replicator.partition.update.count.<device> metrics) but not the
account and container replicators.

This patch extends the existing DB removal count metrics to make them
per-device.  The new metrics are:
 account-replicator.removes.<device>
 container-replicator.removes.<device>

There's also a bonus refactoring and increased test coverage of the DB
replicator code.

Change-Id: Ie46e2c7aa7b091d34476d58a8c86f51fc7ece9d6